applied permissions to users based on their role

+++ resources/views/articles/list.blade.php
@@ -5,7 +5,9 @@
                 {{ __('Articles') }}
             </h2>
 
-            <a href="{{ route('articles.create') }}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">Create</a>
+            @can('create articles')
+                <a href="{{ route('articles.create') }}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">Create</a>
+            @endcan
         </div>
     </x-slot>
 
@@ -42,8 +44,13 @@
                                     {{ \Carbon\Carbon::parse($article->created_at)->format('d M, Y') }}
                                 </td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('articles.edit', $article->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md px-3 py-2 text-white">Edit</a>
-                                    <a href="javascript:void(0);" onclick="deleteArticle({{ $article->id }})" class="bg-red-700 hover:bg-red-600 text-sm rounded-md px-3 py-2 text-white">Delete</a>
+                                    @can('edit articles')
+                                        <a href="{{ route('articles.edit', $article->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md px-3 py-2 text-white">Edit</a>
+                                    @endcan
+                                    
+                                    @can('delete articles')                                        
+                                        <a href="javascript:void(0);" onclick="deleteArticle({{ $article->id }})" class="bg-red-700 hover:bg-red-600 text-sm rounded-md px-3 py-2 text-white">Delete</a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
