+++ resources/views/roles/list.blade.php
@@ -5,7 +5,9 @@
                 {{ __('Roles') }}
             </h2>
 
-            <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">Create</a>
+            @can('create roles')
+                <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white">Create</a>
+            @endcan
         </div>
     </x-slot>
 
@@ -41,8 +43,13 @@
                                     {{ \Carbon\Carbon::parse($role->created_at)->format('d M, Y') }}
                                 </td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('roles.edit', $role->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md px-3 py-2 text-white">Edit</a>
-                                    <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-700 hover:bg-red-600 text-sm rounded-md px-3 py-2 text-white">Delete</a>
+                                    @can('edit roles')
+                                        <a href="{{ route('roles.edit', $role->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md px-3 py-2 text-white">Edit</a>
+                                    @endcan
+
+                                    @can('delete roles')
+                                        <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-700 hover:bg-red-600 text-sm rounded-md px-3 py-2 text-white">Delete</a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
